requirements.txt
numpy
Flask-RESTful
flask-login
flask-sqlalchemy
Flask-PyMongo
Flask-WTF
pandas
matplotlib
mpld3
ipywidgets
